refacto: extract IntentIframe from IntentModal 🔨
