+++ docs/styleguide.config.js
@@ -51,6 +51,7 @@ module.exports = {
       components: () => [
         '../react/Overlay/index.jsx',
         '../react/Alerter/index.jsx',
+        '../react/IntentIframe/IntentIframe.jsx',
         '../react/IntentModal/IntentModal.jsx',
         '../react/IntentOpener/IntentOpener.jsx'
       ]
