+++ react/IntentIframe/styles.styl
@@ -1,0 +1,12 @@
+iframe
+    width 100%
+    height 100%
+    border 0
+
+
+.intentIframe__loading
+    height 100vh
+    width 100vw
+    display flex
+    justify-content center
+    align-items center
