+++ react/IntentModal/styles.styl
@@ -22,18 +22,6 @@
         // we can remove this code
         max-width unset !important
 
-    iframe
-        width 100%
-        height 100%
-        border 0
-
-
-.intentModal__loading
-    height 100%
-    display flex
-    justify-content center
-    align-items center
-
 // TODO remove the second selector when specificity
 // of cozy-ui selectors has decreased
 .intentModal__cross.intentModal__cross
