+++ react/index.js
@@ -6,6 +6,7 @@ export {default as Sidebar} from './Sidebar'
 export {default as Nav, NavItem, NavIcon, NavText, NavLink, genNavLink} from './Nav'
 export {default as Modal, ModalContent, ModalSection, ModalFooter, ModalDescription, ModalHeader, ModalBrandedHeader} from './Modal'
 export {default as SelectionBar} from './SelectionBar'
+export {default as IntentIframe} from './IntentIframe'
 export {default as IntentOpener} from './IntentOpener'
 export {default as IntentModal} from './IntentModal'
 export {default as Spinner} from './Spinner'
